Move control display settings popover to separate file
